Added styling to multiple pages
- Added navigation bar
- Added Login page
- Added Sign-up page
