+++ client/src/Styles/navBar.scss
@@ -1,0 +1,6 @@
+//This file contains all the CSS styling of the navbar
+@import "./variables/variables.scss";
+
+#navbar_title{
+    color: $blue_Color;
+}
