+++ client/src/Styles/variables/variables.scss
@@ -1,0 +1,5 @@
+//This file contains all the variables used in the CSS Styling
+
+$grey_Color: rgb(221, 220, 220);
+$blue_Color: rgb(36, 44, 159);
+
