+++ client/src/components/home.js
@@ -1,10 +1,14 @@
 import React from 'react';
+import NavBar from './navBar';
 function Home() {
 
     return (
-        <div>
-            Home
-        </div>
+        <>
+            <NavBar></NavBar>
+            <div>
+                Home
+            </div>
+        </>
     );
 }
