+++ client/src/index.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import ReactDOM from 'react-dom';
-import './assets/scss/main.scss';
-import Login from "./components/login/login";
-import Signup from "./components/login/signup";
+import Login from "./components/login";
+import Signup from "./components/signup";
 import Home from "./components/home";
+import "bootstrap/dist/css/bootstrap.css"
+
 
 ReactDOM.render(
     <React.StrictMode>
